VictorLiveSessionAspCore/Day4/ExceptionHanldingApp/ExcpetionApp/ExcpetionApp/Controllers/HomeController.cs
using ExcpetionApp.Models;
using Microsoft.AspNetCore.Mvc;
using System.Diagnostics;

namespace ExcpetionApp.Controllers
{
    public class HomeController : Controller
    {
        private readonly ILogger<HomeController> _logger;

        public HomeController(ILogger<HomeController> logger)
        {
            _logger = logger;
            _logger.LogInformation("Home controller created".ToUpper());
        }

        public IActionResult Index()
        {
            return View();
        }

        public IActionResult Privacy()
        {
            _logger.LogError("error in privacy ".ToUpper());
            throw new Exception("some thing is wrong");

            return View();
        }

        [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
        public IActionResult Error()
        {
            return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
        }
    }
}